+++ sdk/cosmos/src/responses/create_stored_procedure_response.rs
@@ -1,4 +1,4 @@
-use crate::from_headers::*;
+use crate::headers::from_headers::*;
 use crate::resources::StoredProcedure;
 use crate::CosmosError;
 use crate::ResourceQuota;
