+++ sdk/cosmos/src/responses/create_trigger_response.rs
@@ -1,4 +1,4 @@
-use crate::from_headers::*;
+use crate::headers::from_headers::*;
 use crate::resources::Trigger;
 use crate::CosmosError;
 use crate::ResourceQuota;
