+++ sdk/cosmos/src/responses/create_user_defined_function_response.rs
@@ -1,4 +1,4 @@
-use crate::from_headers::*;
+use crate::headers::from_headers::*;
 use crate::resources::UserDefinedFunction;
 use crate::CosmosError;
 use crate::ResourceQuota;
