+++ sdk/cosmos/src/responses/delete_stored_procedure_response.rs
@@ -1,4 +1,4 @@
-use crate::from_headers::*;
+use crate::headers::from_headers::*;
 use crate::CosmosError;
 use crate::ResourceQuota;
 use azure_core::headers::session_token_from_headers;
