+++ sdk/cosmos/src/responses/delete_user_response.rs
@@ -1,4 +1,4 @@
-use crate::from_headers::*;
+use crate::headers::from_headers::*;
 use crate::CosmosError;
 use http::response::Response;
