+++ sdk/cosmos/src/responses/get_document_response.rs
@@ -1,4 +1,4 @@
-use crate::from_headers::*;
+use crate::headers::from_headers::*;
 use crate::resources::Document;
 use crate::CosmosError;
 use crate::ResourceQuota;
