+++ sdk/cosmos/src/responses/list_attachments_response.rs
@@ -1,4 +1,4 @@
-use crate::from_headers::*;
+use crate::headers::from_headers::*;
 use crate::resources::Attachment;
 use crate::{CosmosError, ResourceQuota};
 use azure_core::headers::{continuation_token_from_headers_optional, session_token_from_headers};
