+++ sdk/cosmos/src/responses/list_collections_response.rs
@@ -1,4 +1,4 @@
-use crate::from_headers::*;
+use crate::headers::from_headers::*;
 use crate::resources::Collection;
 use crate::CosmosError;
 use crate::ResourceQuota;
