+++ sdk/cosmos/src/responses/list_databases_response.rs
@@ -1,4 +1,4 @@
-use crate::from_headers::*;
+use crate::headers::from_headers::*;
 use crate::resources::Database;
 use crate::CosmosError;
 use crate::ResourceQuota;
