+++ sdk/cosmos/src/responses/list_triggers_response.rs
@@ -1,4 +1,4 @@
-use crate::from_headers::*;
+use crate::headers::from_headers::*;
 use crate::resources::Trigger;
 use crate::CosmosError;
 use crate::ResourceQuota;
