+++ sdk/cosmos/src/responses/list_users_response.rs
@@ -1,4 +1,4 @@
-use crate::from_headers::*;
+use crate::headers::from_headers::*;
 use crate::resources::User;
 use crate::CosmosError;
 use azure_core::headers::{continuation_token_from_headers_optional, session_token_from_headers};
