+++ sdk/cosmos/src/responses/replace_document_response.rs
@@ -1,4 +1,4 @@
-use crate::from_headers::*;
+use crate::headers::from_headers::*;
 use crate::resources::document::DocumentAttributes;
 use crate::CosmosError;
 use crate::ResourceQuota;
